+++ io.openems.edge.ess.sinexcel/bnd.bnd
@@ -3,12 +3,9 @@ Bundle-Vendor: FENECON GmbH
 Bundle-License: https://opensource.org/licenses/EPL-2.0
 Bundle-Version: 1.0.0.${tstamp}
 Export-Package: \
-	io.openems.edge.ess.symmetric.readonly.api,\
-	io.openems.edge.ess.api,\
-	io.openems.edge.ess.symmetric.api,\
-	io.openems.edge.ess.power.symmetric,\
+	io.openems.edge.meter.api,\
 	io.openems.edge.ess.power.api,\
-	io.openems.edge.meter.api
+	io.openems.edge.ess.api
  
  Private-Package: \
  	io.openems.edge.ess.sinexcel
